feat: Correções, testes restruturação, melhorias estruturais e Authentication

+++ foster-pet-user/src/main/java/com/foster/pet/service/impl/AuthenticationServiceImpl.java
@@ -2,43 +2,95 @@ package com.foster.pet.service.impl;
 
 import java.util.Optional;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.foster.pet.constant.ErrorCode;
+import com.foster.pet.dto.authentication.AuthenticationDTO;
+import com.foster.pet.dto.authentication.AuthenticationRDTO;
 import com.foster.pet.entity.Authentication;
+import com.foster.pet.exception.authentication.AuthenticationAlreadyExistsException;
+import com.foster.pet.exception.authentication.AuthenticationNotFoundException;
 import com.foster.pet.repository.AuthenticationRepository;
 import com.foster.pet.service.AuthenticationService;
+import com.foster.pet.util.Encryptor;
 
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 @Service
 public class AuthenticationServiceImpl implements AuthenticationService {
-	private static final Logger log = LoggerFactory.getLogger(AuthenticationServiceImpl.class);
+
+	@Autowired
+	private ModelMapper mapper;
 
 	@Autowired
 	private AuthenticationRepository authenticationRepository;
 
 	@Override
-	public void deleteById(Long id) {
-		log.info("Removendo uma autenticação pelo Id: {}", id);
-		this.authenticationRepository.deleteById(id);
+	public Authentication findById(Long id) {
+		log.info("Start - AuthenticationServiceImpl.findById - Id: {}", id);
+
+		Optional<Authentication> optAuthentication = this.authenticationRepository.findById(id);
+		if (optAuthentication.isEmpty()) {
+			log.error("AuthenticationNotFoundException - Id: {}", id);
+			throw new AuthenticationNotFoundException(ErrorCode.AUTHENTICATION_NOT_FOUND.getMessage());
+		}
+
+		log.info("End - AuthenticationServiceImpl.findById - Authentication: {}", optAuthentication.get().toString());
+		return optAuthentication.get();
 	}
 
 	@Override
-	public Optional<Authentication> findById(Long id) {
-		log.info("Buscando uma autenticação pelo Id: {}", id);
-		return this.authenticationRepository.findById(id);
+	public Authentication findByEmail(String email) {
+		log.info("Start - AuthenticationServiceImpl.findByEmail - Email: {}", email);
+
+		Optional<Authentication> optAuthentication = this.authenticationRepository.findByEmail(email);
+		if (optAuthentication.isEmpty()) {
+			log.error("AuthenticationNotFoundException - Email: {}", email);
+			throw new AuthenticationNotFoundException(ErrorCode.AUTHENTICATION_NOT_FOUND.getMessage());
+		}
+
+		log.info("End - AuthenticationServiceImpl.findByEmail - Authentication: {}",
+				optAuthentication.get().toString());
+		return optAuthentication.get();
 	}
 
 	@Override
-	public Optional<Authentication> findByEmail(String email) {
-		log.info("Buscando uma autenticação pelo email: {}", email);
-		return this.authenticationRepository.findByEmail(email);
-	}
+	public Authentication persist(AuthenticationDTO authenticationDTO) {
+		log.info("Start - AuthenticationServiceImpl.persist - AuthenticationDTO: {}", authenticationDTO.toString());
 
-	@Override
-	public Authentication persist(Authentication authentication) {
-		log.info("Persistindo autenticação: {}", authentication);
+		Authentication authentication = this.mapper.map(authenticationDTO, Authentication.class);
+
+		Optional<Authentication> optAuthentication = this.authenticationRepository
+				.findByEmail(authentication.getEmail());
+		if (optAuthentication.isPresent()) {
+			log.error("AuthenticationAlreadyExistsException - Email: {}", authentication.getEmail());
+			throw new AuthenticationAlreadyExistsException(ErrorCode.AUTHENTICATION_ALREADY_EXISTS.getMessage());
+		}
+
+		String encodedPassword = Encryptor.encode(authentication.getPassword());
+		authentication.setPassword(encodedPassword);
+
+		log.info("End - AuthenticationServiceImpl.persist - Authentication: {}", authentication.toString());
 		return this.authenticationRepository.save(authentication);
+	}
+
+	@Override
+	public AuthenticationRDTO deleteById(Long id) {
+		log.info("Start - AuthenticationServiceImpl.deleteById - Id: {}", id);
+
+		Optional<Authentication> optAuthentication = this.authenticationRepository.findById(id);
+		if (optAuthentication.isEmpty()) {
+			log.error("AuthenticationNotFoundException - Id: {}", id);
+			throw new AuthenticationNotFoundException(ErrorCode.AUTHENTICATION_NOT_FOUND.getMessage());
+		}
+
+		this.authenticationRepository.deleteById(id);
+		AuthenticationRDTO authenticationDTO = this.mapper.map(optAuthentication.get(), AuthenticationRDTO.class);
+
+		log.info("End - AuthenticationServiceImpl.deleteById - AuthenticationDTO: {}", authenticationDTO.toString());
+		return authenticationDTO;
 	}
 }
